requirements.txt
Flask==2.3.2
google-api-python-client==2.88.0
google-auth-oauthlib==1.0.0
requests==2.31.0
